backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic>=2.7.0
pydantic[email]>=2.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
apscheduler>=3.10.0
httpx>=0.27.0
openai>=1.0.0
python-multipart>=0.0.9
alembic>=1.13.0
psycopg2-binary>=2.9.9
